+++ tests/scalers/azure/azure_queue_aad_wi/azure_queue_aad_wi_test.go
@@ -162,11 +162,14 @@ func getTemplateData() (templateData, []Template) {
 			SecretName:       secretName,
 			Connection:       base64ConnectionString,
 			DeploymentName:   deploymentName,
+			TriggerAuthName:  triggerAuthName,
 			ScaledObjectName: scaledObjectName,
+			AccountName:      accountName,
 			QueueName:        queueName,
 		}, []Template{
 			{Name: "secretTemplate", Config: secretTemplate},
 			{Name: "deploymentTemplate", Config: deploymentTemplate},
+			{Name: "triggerAuthTemplate", Config: triggerAuthTemplate},
 			{Name: "scaledObjectTemplate", Config: scaledObjectTemplate},
 		}
 }