+++ CHANGELOG.md
@@ -168,6 +168,9 @@ New deprecation(s):
 - **General**: Minor refactor to reduce copy/paste code in ScaledObject webhook ([#5397](https://github.com/kedacore/keda/issues/5397))
 - **General**: No need to list all secret in the namespace to find just one ([#5669](https://github.com/kedacore/keda/pull/5669))
 - **Kafka**: Expose GSSAPI service name  ([#5474](https://github.com/kedacore/keda/issues/5474))
+- **Azure Event Hub Scaler**: Migrated to `github.com/Azure/azure-sdk-for-go/sdk/messaging/azeventhubs` ([#5470](https://github.com/kedacore/keda/issues/5470))
+- **Azure Queue Scaler**: Migrated to `github.com/Azure/azure-sdk-for-go/sdk/storage/azqueue` ([#5470](https://github.com/kedacore/keda/issues/5470))
+- **Azure Storage Scaler**: Migrated to `github.com/Azure/azure-sdk-for-go/sdk/storage/azblob` ([#5470](https://github.com/kedacore/keda/issues/5470))
 
 ## v2.13.1